Finish moving components to create LoginContainer
